requirements.txt
fastapi
uvicorn
openai
scikit-learn
pydantic
numpy
python-dotenv